Ability to override startup checks

+++ index.php
@@ -10,17 +10,17 @@ require("includes/settings.php");
 require("functions/cmum.php");
 require("functions/login.php");
 
-if(file_exists("install/")) {
+if(file_exists("install/") && !isset($overridechecks)) {
 	errorpage("INSTALLATION INCOMPLETE","Please delete the install dir from your www-root after you completed the installation!","utf-8",CMUM_TITLE,$_SERVER["REQUEST_URI"],CMUM_VERSION,CMUM_BUILD,CMUM_MOD);
 	exit;
 }
 
-if(checkversion(CMUM_VERSION)=="0") {
+if(checkversion(CMUM_VERSION)=="0" && !isset($overridechecks)) {
 	errorpage("VERSION MISMATCH","Please run the upgrade tool to upgrade your installation!<br><br>Click <a href=\"\upgrade\">here</a> to run the upgrade tool","utf-8",CMUM_TITLE,$_SERVER["REQUEST_URI"],CMUM_VERSION,CMUM_BUILD,CMUM_MOD);
 	exit;
 }
 
-if(file_exists("upgrade/")) {
+if(file_exists("upgrade/") && !isset($overridechecks)) {
 	errorpage("UPGRADE INCOMPLETE","Please delete the upgrade dir from your www-root after you completed the upgrade!","utf-8",CMUM_TITLE,$_SERVER["REQUEST_URI"],CMUM_VERSION,CMUM_BUILD,CMUM_MOD);
 	exit;
 }
